fixed the compile warnings */

+++ raymarch.cpp
@@ -4,8 +4,9 @@
 
 #include "color.hpp"
 
-const unsigned long int W = 1318;
-const unsigned long int H = 716;
+/* window size.  Seems 'uint' is fine type here */
+const uint W = 1318;
+const uint H = 716;
 
 //bool world[W][H];
 //bool buffer[W][H];
@@ -62,8 +63,8 @@ double clamp(double n,double min,double max) {
 
 void render(void) {
     Color color;
-    for(long int x = 0; x < W; x++) {
-        for(long int y = 0; y < H; y++) {
+    for(uint x = 0; x < W; x++) {
+        for(uint y = 0; y < H; y++) {
             double tx = ((double)x-W/2)/3;
             double ty = ((double)y-H/2)/3;
             sf::Vector3f dir(tx*0.001,ty*0.001,1);
@@ -91,8 +92,8 @@ int main()
 
     sprite.setTexture(texture);
 
-    for(int x = 0; x < W; x++) {
-        for(int y = 0; y < H; y++) {
+    for(uint x = 0; x < W; x++) {
+        for(uint y = 0; y < H; y++) {
             pixels[3+x*4+y*4*W] = 255;
         }
     }
